+++ R/slide.R
@@ -32,7 +32,7 @@ replace_fn_names <- function(fn, replace = list()){
 #' @param .align Align index at the "**r**ight", "**c**entre"/"center", or "**l**eft"
 #' of the window. If `.size` is even for center alignment, "centre-right" & "centre-left"
 #' is needed.
-#' @param .combine Should `.x` be combined before applying `.f`.
+#' @param .bind If `.x` is a list, should `.x` be combined before applying `.f`?
 #'
 #' @rdname slide
 #' @export
@@ -54,11 +54,11 @@ replace_fn_names <- function(fn, replace = list()){
 #' slide(lst, ~ ., .size = 2)
 slide <- function(
   .x, .f, ..., .size = 1, .fill = NA, .partial = FALSE, 
-  .align = "right", .combine = FALSE
+  .align = "right", .bind = FALSE
 ) {
   lst_x <- slider(
     .x, .size = .size, .fill = .fill, .partial = .partial, 
-    .align = .align, .combine = .combine
+    .align = .align, .bind = .bind
   )
   out <- purrr::map(lst_x, .f, ...)
   if (.partial) return(out)
@@ -80,12 +80,12 @@ for(type in c("lgl", "chr", "int", "dbl")){
 #' @export
 slide_dfr <- function(
   .x, .f, ..., .size = 1, .fill = NA, .partial = FALSE, .align = "right", 
-  .combine = FALSE, .id = NULL
+  .bind = FALSE, .id = NULL
 ) {
   out <- slide(
     .x, .f = .f, ..., 
     .size = .size, .fill = .fill, .partial = .partial, 
-    .align = .align, .combine = .combine
+    .align = .align, .bind = .bind
   )
   bind_df(out, .size, .fill, .id = .id)
 }
@@ -94,12 +94,12 @@ slide_dfr <- function(
 #' @export
 slide_dfc <- function(
   .x, .f, ..., .size = 1, .fill = NA, .partial = FALSE, 
-  .align = "right", .combine = FALSE
+  .align = "right", .bind = FALSE
 ) {
   out <- slide(
     .x, .f = .f, ..., 
     .size = .size, .fill = .fill, .partial = .partial, 
-    .align = .align, .combine = .combine
+    .align = .align, .bind = .bind
   )
   bind_df(out, .size, .fill, byrow = FALSE)
 }
@@ -136,11 +136,11 @@ slide_dfc <- function(
 #' pslide(list(lst, lst), ~ ., .size = 2)
 slide2 <- function(
   .x, .y, .f, ..., .size = 1, .fill = NA, .partial = FALSE, 
-  .align = "right", .combine = FALSE
+  .align = "right", .bind = FALSE
 ) {
   lst <- pslider(
     .x, .y, .size = .size, .fill = .fill, .partial = .partial, 
-    .align = .align, .combine = .combine
+    .align = .align, .bind = .bind
   )
   out <- purrr::map2(lst[[1]], lst[[2]], .f = .f, ...)
   if (.partial) return(out)
@@ -162,12 +162,12 @@ for(type in c("lgl", "chr", "int", "dbl")){
 #' @export
 slide2_dfr <- function(
   .x, .y, .f, ..., .size = 1, .fill = NA, .partial = FALSE, .align = "right",
-  .combine = FALSE, .id = NULL
+  .bind = FALSE, .id = NULL
 ) {
   out <- slide2(
     .x, .y, .f = .f, ..., 
     .size = .size, .fill = .fill, .partial = .partial, 
-    .align = .align, .combine = .combine
+    .align = .align, .bind = .bind
   )
   bind_df(out, .size, .fill, .id = .id)
 }
@@ -176,12 +176,12 @@ slide2_dfr <- function(
 #' @export
 slide2_dfc <- function(
   .x, .y, .f, ..., .size = 1, .fill = NA, .partial = FALSE, 
-  .align = "right", .combine = FALSE
+  .align = "right", .bind = FALSE
 ) {
   out <- slide2(
     .x, .y, .f = .f, ..., 
     .size = .size, .fill = .fill, .partial = .partial, 
-    .align = .align, .combine = .combine
+    .align = .align, .bind = .bind
   )
   bind_df(out, .size, .fill, byrow = FALSE)
 }
@@ -191,11 +191,11 @@ slide2_dfc <- function(
 #' @export
 pslide <- function(
   .l, .f, ..., .size = 1, .fill = NA, .partial = FALSE, 
-  .align = "right", .combine = FALSE
+  .align = "right", .bind = FALSE
 ) {
   lst <- pslider(
     !!! .l, .size = .size, .fill = .fill, .partial = .partial, 
-    .align = .align, .combine = .combine
+    .align = .align, .bind = .bind
   )
   out <- purrr::pmap(lst, .f, ...)
   if (.partial) return(out)
@@ -217,12 +217,12 @@ for(type in c("lgl", "chr", "int", "dbl")){
 #' @export
 pslide_dfr <- function(
   .l, .f, ..., .size = 1, .fill = NA, .partial = FALSE, .align = "right",
-  .combine = FALSE, .id = NULL
+  .bind = FALSE, .id = NULL
 ) {
   out <- pslide(
     .l, .f = .f, ..., 
     .size = .size, .fill = .fill, .partial = .partial, 
-    .align = .align, .combine = .combine
+    .align = .align, .bind = .bind
   )
   bind_df(out, .size, .fill, .id = .id)
 }
@@ -250,12 +250,12 @@ pslide_dfr <- function(
 #' }
 pslide_dfc <- function(
   .l, .f, ..., .size = 1, .fill = NA, .partial = FALSE, 
-  .align = "right", .combine = FALSE
+  .align = "right", .bind = FALSE
 ) {
   out <- pslide(
     .l, .f = .f, ..., 
     .size = .size, .fill = .fill, .partial = .partial, 
-    .align = .align, .combine = .combine
+    .align = .align, .bind = .bind
   )
   bind_df(out, .size, .fill, byrow = FALSE)
 }
@@ -266,7 +266,7 @@ pslide_dfc <- function(
 #' @param ... Multiple objects to be split in parallel.
 #' @param .partial if `TRUE`, split to partial set (`FALSE` ignores specified 
 #' `.fill` and `.align`).
-#' @param .combine If `.x` is a list or data frame, the input will be flattened
+#' @param .bind If `.x` is a list or data frame, the input will be flattened
 #' to a list of data frames.
 #' @inheritParams slide
 #' @rdname slider
@@ -280,16 +280,15 @@ pslide_dfc <- function(
 #'
 #' slider(x, .size = 2)
 #' slider(lst, .size = 2)
-#' slider(df, .size = 2, .combine = TRUE)
 #' pslider(list(x, y), list(y))
 #' slider(df, .size = 2)
 #' pslider(df, df, .size = 2)
 slider <- function(
   .x, .size = 1, .fill = NA, .partial = FALSE, 
-  .align = "right", .combine = FALSE
+  .align = "right", .bind = FALSE
 ) {
   bad_window_function(.size)
-  .x <- df2lst(.x, .combine)
+  if (is.data.frame(.x)) .x <- as.list(.x)
   len_x <- NROW(.x)
   abs_size <- abs(.size)
   if (abs_size > len_x) {
@@ -299,45 +298,46 @@ slider <- function(
     ))
   }
   sign <- sign(.size)
-  is_df <- is.data.frame(.x)
   if (.partial) {
     lst_idx <- seq_len(len_x) - sign * (abs_size - 1)
     if (sign < 0) lst_idx <- rev(lst_idx)
     return(purrr::map(lst_idx, function(idx) {
       idx <- idx:(idx + sign * (abs_size - 1))
       size <- sum(idx <= 0 | idx > len_x) + 1
-      if (is_df) 
-        return(pad_slide(
-          .x[idx[idx > 0 & idx <= len_x], , drop = FALSE], size, .fill, .align
-        ))
       out <- pad_slide(.x[idx[idx > 0 & idx <= len_x]], size, .fill, .align)
-      if (is_bare_list(.x) && .combine) return(lapply(out, dplyr::bind_rows))
-      return(out)
+      if (.bind) return(bind_lst(out)) else return(out)
     }))
   }
   lst_idx <- seq_len(len_x - abs_size + 1)
   if (sign < 0) lst_idx <- rev(lst_idx) + 1
-  if (is_df)
-    return(purrr::map(lst_idx, 
-      function(idx) .x[idx:(idx + sign * (abs_size - 1)), , drop = FALSE]
-    ))
   out <- purrr::map(lst_idx, function(idx) .x[idx:(idx + sign * (abs_size - 1))])
-  if (is_bare_list(.x) && .combine) {
-    return(lapply(out, dplyr::bind_rows))
-  }
-  out
+  if (.bind) return(bind_lst(out)) else out
 }
 
 #' @rdname slider
 #' @export
 pslider <- function(
   ..., .size = 1, .fill = NA, .partial = FALSE, 
-  .align = "right", .combine = FALSE
+  .align = "right", .bind = FALSE
 ) { # parallel sliding
   lst <- recycle(list2(...))
   purrr::map(lst, 
-    function(x) slider(x, .size, .fill = .fill, .partial, .align, .combine)
+    function(x) slider(x, .size, .fill = .fill, .partial, .align, .bind)
   )
+}
+
+bind_lst <- function(x) {
+  if (!is.list(x)) {
+    abort(sprintf("`.bind = TRUE` only accepts a list, not %s.", typeof(x)))
+  }
+  cls_elements <- flatten_chr(purrr::modify_depth(x, 2, ~ class(.)[1]))
+  cls <- purrr::reduce(cls_elements, identical_class)
+  if (cls == "data.frame") {
+    out <- lapply(x, dplyr::bind_rows)
+  } else {
+    out <- lapply(x, function(y) do.call(c, y))
+  }
+  out
 }
 
 bad_window_function <- function(.size) {
@@ -411,18 +411,4 @@ check_valid_window <- function(.size, .align) {
       .align
     ))
   }
-}
-
-df2lst <- function(x, .combine = FALSE) {
-  if (.combine && is_atomic(x)) 
-    warn(sprintf("`.combine = TRUE` is ignored for type `%s`.", typeof(x)))
-  if (is_false(.combine)) {
-    if (is.data.frame(x)) {
-      x <- as.list(x)
-    } else if (is_bare_list(x)){
-      df_lgl <- purrr::map_lgl(x, is.data.frame)
-      if (any(df_lgl)) x[df_lgl] <- purrr::map(x[df_lgl], as.list)
-    }
-  }
-  x
 }