+++ R/stretch.R
@@ -22,8 +22,8 @@
 #' stretch_dbl(x, mean, .size = 2)
 #' stretch_lgl(x, ~ mean(.) > 2, .size = 2)
 #' stretch(lst, ~ ., .size = 2)
-stretch <- function(.x, .f, ..., .size = 1, .init = 1, .combine = FALSE) {
-  lst_x <- stretcher(.x, .size = .size, .init = .init, .combine)
+stretch <- function(.x, .f, ..., .size = 1, .init = 1, .bind = FALSE) {
+  lst_x <- stretcher(.x, .size = .size, .init = .init, .bind)
   purrr::map(lst_x, .f, ...)
 }
 
@@ -41,21 +41,21 @@ for(type in c("lgl", "chr", "dbl", "int")){
 #' @rdname stretch
 #' @export
 stretch_dfr <- function(
-  .x, .f, ..., .size = 1, .init = 1, .combine = FALSE, .id = NULL
+  .x, .f, ..., .size = 1, .init = 1, .bind = FALSE, .id = NULL
 ) {
   out <- stretch(
     .x, .f = .f, ..., .size = .size, .init = .init, 
-    .combine = .combine
+    .bind = .bind
   )
   dplyr::bind_rows(!!! out, .id = .id)
 }
 
 #' @rdname stretch
 #' @export
-stretch_dfc <- function(.x, .f, ..., .size = 1, .init = 1, .combine = FALSE) {
+stretch_dfc <- function(.x, .f, ..., .size = 1, .init = 1, .bind = FALSE) {
   out <- stretch(
     .x, .f = .f, ..., .size = .size, .init = .init,
-    .combine = .combine
+    .bind = .bind
   )
   dplyr::bind_cols(!!! out)
 }
@@ -89,8 +89,8 @@ stretch_dfc <- function(.x, .f, ..., .size = 1, .init = 1, .combine = FALSE) {
 #' stretch2(df, df, ~ ., .size = 2)
 #' pstretch(lst, sum, .size = 1)
 #' pstretch(list(lst, lst), ~ ., .size = 2)
-stretch2 <- function(.x, .y, .f, ..., .size = 1, .init = 1, .combine = FALSE) {
-  lst <- pstretcher(.x, .y, .size = .size, .init = .init, .combine = .combine)
+stretch2 <- function(.x, .y, .f, ..., .size = 1, .init = 1, .bind = FALSE) {
+  lst <- pstretcher(.x, .y, .size = .size, .init = .init, .bind = .bind)
   purrr::map2(lst[[1]], lst[[2]], .f, ...)
 }
 
@@ -108,11 +108,11 @@ for(type in c("lgl", "chr", "dbl", "int")){
 #' @rdname stretch2
 #' @export
 stretch2_dfr <- function(
-  .x, .y, .f, ..., .size = 1, .init = 1, .combine = FALSE, .id = NULL
+  .x, .y, .f, ..., .size = 1, .init = 1, .bind = FALSE, .id = NULL
 ) {
   out <- stretch2(
     .x, .y, .f = .f, ..., .size = .size, .init = .init,
-    .combine = .combine
+    .bind = .bind
   )
   dplyr::bind_rows(!!! out, .id = .id)
 }
@@ -120,19 +120,19 @@ stretch2_dfr <- function(
 #' @rdname stretch2
 #' @export
 stretch2_dfc <- function(
-  .x, .y, .f, ..., .size = 1, .init = 1, .combine = FALSE
+  .x, .y, .f, ..., .size = 1, .init = 1, .bind = FALSE
 ) {
   out <- stretch2(
     .x, .y, .f = .f, ..., .size = .size, .init = .init,
-    .combine = .combine
+    .bind = .bind
   )
   dplyr::bind_cols(!!! out)
 }
 
 #' @rdname stretch2
 #' @export
-pstretch <- function(.l, .f, ..., .size = 1, .init = 1, .combine = FALSE) {
-  lst <- pstretcher(!!! .l, .size = .size, .init = .init, .combine = .combine)
+pstretch <- function(.l, .f, ..., .size = 1, .init = 1, .bind = FALSE) {
+  lst <- pstretcher(!!! .l, .size = .size, .init = .init, .bind = .bind)
   purrr::pmap(lst, .f, ...)
 }
 
@@ -150,16 +150,16 @@ for(type in c("lgl", "chr", "dbl", "int")){
 #' @rdname stretch2
 #' @export
 pstretch_dfr <- function(
-  .l, .f, ..., .size = 1, .init = 1, .combine = FALSE, .id = NULL
+  .l, .f, ..., .size = 1, .init = 1, .bind = FALSE, .id = NULL
 ) {
-  out <- pstretch(.l, .f, ..., .size = .size, .init = .init, .combine = .combine)
+  out <- pstretch(.l, .f, ..., .size = .size, .init = .init, .bind = .bind)
   dplyr::bind_rows(!!! out, .id = .id)
 }
 
 #' @rdname stretch2
 #' @export
-pstretch_dfc <- function(.l, .f, ..., .size = 1, .init = 1, .combine = FALSE) {
-  out <- pstretch(.l, .f, ..., .size = .size, .init = .init, .combine = .combine)
+pstretch_dfc <- function(.l, .f, ..., .size = 1, .init = 1, .bind = FALSE) {
+  out <- pstretch(.l, .f, ..., .size = .size, .init = .init, .bind = .bind)
   dplyr::bind_cols(!!! out)
 }
 
@@ -179,15 +179,14 @@ pstretch_dfc <- function(.l, .f, ..., .size = 1, .init = 1, .combine = FALSE) {
 #'
 #' stretcher(x, .size = 2)
 #' stretcher(lst, .size = 2)
-#' stretcher(df, .size = 2, .combine = TRUE)
 #' stretcher(df, .size = 2)
 #' pstretcher(df, df, .size = 2)
-stretcher <- function(.x, .size = 1, .init = 1, .combine = FALSE) {
+stretcher <- function(.x, .size = 1, .init = 1, .bind = FALSE) {
   bad_window_function(.size)
   if (!is_integerish(.init, n = 1) || .init < 1) {
     abort("`.init` must be a positive integer.")
   }
-  .x <- df2lst(.x, .combine)
+  if (is.data.frame(.x)) .x <- as.list(.x)
   len_x <- NROW(.x)
   abs_size <- abs(.size)
   counter <- incr(init = .init, size = abs_size)
@@ -198,21 +197,17 @@ stretcher <- function(.x, .size = 1, .init = 1, .combine = FALSE) {
     purrr::map(ncall, ~ seq_len(counter())),
     list(seq_len(len_x))
   )
-  is_df <- is.data.frame(.x)
-  if (is_df) return(purrr::map(incr_lst, function(idx) .x[idx, , drop = FALSE]))
   out <- purrr::map(incr_lst, function(idx) .x[idx])
-  if (is_bare_list(.x) && .combine) { # a list of data frames
-    return(lapply(out, dplyr::bind_rows))
-  }
+  if (.bind) return(bind_lst(out)) else out
   out
 }
 
 
 #' @rdname stretcher
 #' @export
-pstretcher <- function(..., .size = 1, .init = 1, .combine = FALSE) { # parallel sliding
+pstretcher <- function(..., .size = 1, .init = 1, .bind = FALSE) { # parallel sliding
   lst <- recycle(list2(...))
-  purrr::map(lst, function(x) stretcher(x, .size, .init, .combine))
+  purrr::map(lst, function(x) stretcher(x, .size, .init, .bind))
 }
 
 incr <- function(init, size) {